refactor: Page 레이어 구조 변경
