+++ client/src/pages/admin/ui/admin-page.tsx
@@ -1,5 +1,5 @@
 import { Outlet } from "react-router";
-import { AdminLayout } from "./ui/admin-layout";
+import { AdminLayout } from "./admin-layout";
 
 export function AdminPage() {
   return (
