+++ client/src/pages/counter-selector/index.ts
@@ -1,1 +1,1 @@
-export { CounterSelectorPage } from "./ui/page";
+export { CounterSelectorPage } from "./ui/counter-selector-page";
