+++ client/src/pages/home/index.ts
@@ -1,1 +1,1 @@
-export { HomePage } from "./ui";
+export { HomePage } from "./ui/home-page";
