+++ client/src/pages/manual-counter/ui/manual-counter-page.tsx
